+++ middleware/headers/headers.go
@@ -1,10 +1,10 @@
 package headers
 
 import (
+	"fmt"
 	"net/http"
 
-	"github.com/dusted-go/http/v3/middleware"
-	"github.com/dusted-go/http/v3/response"
+	"github.com/dusted-go/http/v4/middleware"
 )
 
 // Security will set HTTP security headers.
@@ -15,7 +15,12 @@ import (
 func Security(hstsMaxAge int) middleware.Middleware {
 	return middleware.Func(
 		func(next http.Handler, w http.ResponseWriter, r *http.Request) {
-			response.SecurityHeaders(w, hstsMaxAge)
+			w.Header().Set(
+				"Strict-Transport-Security",
+				fmt.Sprintf("max-age=%d; includeSubDomains", hstsMaxAge))
+			w.Header().Set("X-Content-Type-Options", "nosniff")
+			w.Header().Set("X-Frame-Options", "SAMEORIGIN")
+			w.Header().Set("Referrer-Policy", "strict-origin-when-cross-origin")
 			next.ServeHTTP(w, r)
 		},
 	)