+++ route/route.go
@@ -1,10 +1,10 @@
-package path
+package route
 
 import "strings"
 
-// Shift splits off the first component of path.
+// ShiftPath splits off the first component of path.
 // Head will never contain a slash and tail will always be a rooted path without trailing slash.
-func Shift(path string) (head, tail string) {
+func ShiftPath(path string) (head, tail string) {
 	i := strings.Index(path[1:], "/") + 1
 	if i <= 0 {
 		return path[1:], "/"